Extract Monad & Error modules from Blockfrost
