+++ lib/shelley/cardano-wallet.cabal
@@ -120,7 +120,9 @@ library
       Cardano.Wallet.Shelley.Compatibility.Ledger
       Cardano.Wallet.Shelley.Network
       Cardano.Wallet.Shelley.Network.Blockfrost
+      Cardano.Wallet.Shelley.Network.Blockfrost.Error
       Cardano.Wallet.Shelley.Network.Blockfrost.Fixture
+      Cardano.Wallet.Shelley.Network.Blockfrost.Monad
       Cardano.Wallet.Shelley.Network.Discriminant
       Cardano.Wallet.Shelley.Network.Node
       Cardano.Wallet.Shelley.Transaction
