+++ lib/shelley/src/Cardano/Wallet/Shelley/Network/Blockfrost.hs
@@ -1,12 +1,10 @@
 {-# LANGUAGE ApplicativeDo #-}
 {-# LANGUAGE BlockArguments #-}
 {-# LANGUAGE DataKinds #-}
-{-# LANGUAGE DeriveAnyClass #-}
 {-# LANGUAGE DerivingStrategies #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE GADTs #-}
-{-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE NamedFieldPuns #-}
@@ -29,11 +27,7 @@ module Cardano.Wallet.Shelley.Network.Blockfrost
 
     -- * Internal
     getPoolPerformanceEstimate,
-    eraByEpoch,
-    newClientConfig,
-    BFM (..),
-    runBFM,
-    BlockfrostError (..),
+    eraByEpoch
   )
 where
 
@@ -138,28 +132,26 @@ import Cardano.Wallet.Primitive.Types.TokenBundle
     ( TokenBundle (..) )
 import Cardano.Wallet.Primitive.Types.Tx
     ( Tx (..), TxIn (..), TxOut (..), TxScriptValidity (..), TxSize (..) )
+import Cardano.Wallet.Shelley.Network.Blockfrost.Error
+    ( BlockfrostError (..) )
+import Cardano.Wallet.Shelley.Network.Blockfrost.Monad
+    ( BFM )
 import Cardano.Wallet.Shelley.Network.Discriminant
     ( SomeNetworkDiscriminant (..), networkDiscriminantToId )
 import Control.Concurrent
     ( threadDelay )
 import Control.Concurrent.Async.Lifted
     ( concurrently, forConcurrently, mapConcurrently )
-import Control.Exception
-    ( throwIO )
 import Control.Monad
-    ( forever, join, unless, (<=<) )
-import Control.Monad.Base
-    ( MonadBase )
+    ( forever, join, unless )
 import Control.Monad.Error.Class
-    ( MonadError (catchError), liftEither, throwError )
+    ( MonadError, liftEither, throwError )
 import Control.Monad.IO.Class
     ( MonadIO (liftIO) )
-import Control.Monad.Reader
-    ( MonadReader, ReaderT (runReaderT), ask, asks )
 import Control.Monad.Trans.Control
     ( MonadBaseControl )
 import Control.Monad.Trans.Except
-    ( ExceptT (..), runExceptT )
+    ( ExceptT (..) )
 import Data.Align
     ( align )
 import Data.Bifunctor
@@ -202,7 +194,7 @@ import Data.Set
 import Data.Text
     ( Text )
 import Data.Text.Class
-    ( FromText (fromText), TextDecodingError (..), ToText (..) )
+    ( FromText (fromText), ToText (..) )
 import Data.These
     ( These (That, These, This) )
 import Data.Traversable
@@ -213,22 +205,16 @@ import GHC.OldList
     ( sortOn )
 import Money
     ( Discrete' )
-import Network.HTTP.Types
-    ( status404 )
 import Ouroboros.Consensus.Cardano.Block
     ( CardanoBlock, StandardCrypto )
 import Ouroboros.Consensus.HardFork.History.EraParams
     ( EraParams (..), SafeZone (StandardSafeZone, UnsafeIndefiniteSafeZone) )
 import Ouroboros.Consensus.HardFork.History.Summary
     ( Bound (..), EraEnd (..), EraSummary (..), Summary (..) )
-import Servant.Client
-    ( runClientM )
 import Text.Read
     ( readEither )
 import UnliftIO.Async
     ( async, link )
-import UnliftIO.Exception
-    ( Exception )
 import UnliftIO.STM
     ( TChan
     , atomically
@@ -245,6 +231,7 @@ import qualified Cardano.Slotting.Time as ST
 import qualified Cardano.Wallet.Network.Light as LN
 import qualified Cardano.Wallet.Primitive.Types.Coin as Coin
 import qualified Cardano.Wallet.Shelley.Network.Blockfrost.Fixture as Fixture
+import qualified Cardano.Wallet.Shelley.Network.Blockfrost.Monad as BFM
 import qualified Data.Aeson as Aeson
 import qualified Data.HashMap.Strict as HashMap
 import qualified Data.Map.Strict as Map
@@ -254,32 +241,10 @@ import qualified Data.Text as T
 import qualified Data.Vector as V
 import qualified Ouroboros.Consensus.HardFork.History.Qry as HF
 import qualified Ouroboros.Consensus.Util.Counting as UC
-import qualified Servant.Client as Servant
 
 {-------------------------------------------------------------------------------
     NetworkLayer
 -------------------------------------------------------------------------------}
-
-data BlockfrostError
-    = ClientError Servant.ClientError
-    | NoSlotError BF.Block
-    | IntegralCastError String
-    | InvalidBlockHash BF.BlockHash TextDecodingError
-    | InvalidTxMetadataLabel String
-    | InvalidTxMetadataValue String
-    | InvalidTxHash Text TextDecodingError
-    | InvalidAddress Text TextDecodingError
-    | InvalidPoolId Text TextDecodingError
-    | PoolStakePercentageError Coin Coin
-    | InvalidDecentralizationLevelPercentage Double
-    | InvalidUtxoInputAmount BF.UtxoInput
-    | InvalidUtxoOutputAmount BF.UtxoOutput
-    | UnknownEraForEpoch EpochNo
-    deriving (Show, Eq)
-
-newtype BlockfrostException = BlockfrostException BlockfrostError
-    deriving stock (Show)
-    deriving anyclass (Exception)
 
 data Log
     = MsgTipReceived BlockHeader
@@ -387,7 +352,7 @@ withNetworkLayer
     -> (NetworkLayer IO (CardanoBlock StandardCrypto) -> IO a)
     -> IO a
 withNetworkLayer tr network np project k = do
-    bfConfig <- newClientConfig project
+    bfConfig <- BFM.newClientConfig project
     tipBroadcast <- newBroadcastTChanIO
     link =<< async (pollNodeTip bfConfig tipBroadcast)
     k NetworkLayer
@@ -416,7 +381,7 @@ withNetworkLayer tr network np project k = do
 
     currentNodeTip :: BF.ClientConfig -> IO BlockHeader
     currentNodeTip bfConfig = do
-        tip <- runBFM bfConfig $ fromBlockfrostM =<< BF.getLatestBlock
+        tip <- BFM.run bfConfig $ fromBlockfrostM =<< BF.getLatestBlock
         traceWith tr $ MsgFetchedLatestBlockHeader tip
         pure tip
 
@@ -424,7 +389,7 @@ withNetworkLayer tr network np project k = do
     pollNodeTip bfConfig nodeTip = do
         lastTip <- atomically $ dupTChan nodeTip
         link =<< async =<< forever do
-            header <- runBFM bfConfig $ fromBlockfrostM =<< BF.getLatestBlock
+            header <- BFM.run bfConfig $ fromBlockfrostM =<< BF.getLatestBlock
             atomically do
                 lastHeader <- tryReadTChan lastTip
                 unless (lastHeader == Just header) do
@@ -441,11 +406,11 @@ withNetworkLayer tr network np project k = do
 
     currentProtocolParameters :: BF.ClientConfig -> IO ProtocolParameters
     currentProtocolParameters bfConfig =
-        runBFM bfConfig $ liftEither . fromBlockfrostPP networkId
+        BFM.run bfConfig $ liftEither . fromBlockfrostPP networkId
             =<< BF.getLatestEpochProtocolParams
 
     currentSlottingParameters :: BF.ClientConfig -> IO SlottingParameters
-    currentSlottingParameters bfConfig = runBFM bfConfig do
+    currentSlottingParameters bfConfig = BFM.run bfConfig do
         liftIO $ traceWith tr MsgCurrentSlottingParameters
         BF.EpochInfo{_epochInfoEpoch} <- BF.getLatestEpoch
         epochNo <- fromBlockfrostM _epochInfoEpoch
@@ -466,7 +431,7 @@ withNetworkLayer tr network np project k = do
 
     currentNodeEra :: BF.ClientConfig -> IO AnyCardanoEra
     currentNodeEra bfConfig = do
-        (era, epoch) <- runBFM bfConfig do
+        (era, epoch) <- BFM.run bfConfig do
             BF.EpochInfo{_epochInfoEpoch} <- BF.getLatestEpoch
             latestEpoch <- fromBlockfrostM _epochInfoEpoch
             latestEra <- liftEither $ eraByEpoch networkId latestEpoch
@@ -488,10 +453,10 @@ withNetworkLayer tr network np project k = do
     fetchNetworkRewardAccountBalances
         (SomeNetworkDiscriminant (Proxy :: Proxy nd)) bfConfig accounts = do
             traceWith tr MsgFetchNetworkRewardAccountBalances
-            runBFM bfConfig $ Map.fromList . catMaybes <$>
+            BFM.run bfConfig $ Map.fromList . catMaybes <$>
                 for (Set.toList accounts) \rewardAccount -> do
                     let addr = BF.mkAddress $ encodeStakeAddress @nd rewardAccount
-                    maybe404 (BF.getAccount addr)
+                    BFM.maybe404 (BF.getAccount addr)
                         >>= traverse \BF.AccountInfo{..} ->
                             (rewardAccount,) . Coin <$>
                                 fromIntegral @_ @Integer _accountInfoRewardsSum
@@ -517,7 +482,7 @@ withNetworkLayer tr network np project k = do
         bfConfig
         follower = do
             let runBF :: forall a. BFM a -> IO a
-                runBF = runBFM bfConfig
+                runBF = BFM.run bfConfig
             AnyCardanoEra era <- currentNodeEra bfConfig
             let sp = slottingParameters np
             let stabilityWindow =
@@ -580,7 +545,7 @@ withNetworkLayer tr network np project k = do
                     runBF $ fromBlockEvents <$> case addrOrAcc of
                         Left address -> do
                             txs <- BF.allPages \paged ->
-                                empty404 $ BF.getAddressTransactions'
+                                BFM.empty404 $ BF.getAddressTransactions'
                                     (BF.Address (encodeAddress @nd address))
                                     paged
                                     BF.Ascending
@@ -600,10 +565,10 @@ withNetworkLayer tr network np project k = do
                             let address = BF.Address $ encodeStakeAddress @nd account
                             regTxHashes <-
                                 fmap BF._accountRegistrationTxHash
-                                    <$> empty404 (BF.getAccountRegistrations address)
+                                    <$> BFM.empty404 (BF.getAccountRegistrations address)
                             delTxHashes <-
                                 fmap BF._accountDelegationTxHash
-                                    <$> empty404 (BF.getAccountDelegations address)
+                                    <$> BFM.empty404 (BF.getAccountDelegations address)
                             blockEventsRegDeleg <-
                                 forConcurrently (regTxHashes <> delTxHashes) \hash -> do
                                     (tx@BF.Transaction{_transactionIndex}, dcerts) <-
@@ -618,7 +583,7 @@ withNetworkLayer tr network np project k = do
                                             (\(n, dc) -> ((txIndex, n), dc))
                                                 <$> zip [0 ..] dcerts
                                         )
-                            ws <- empty404 (BF.getAccountWithdrawals address)
+                            ws <- BFM.empty404 (BF.getAccountWithdrawals address)
                             blockEventsWithdraw <-
                                 forConcurrently ws \BF.AccountWithdrawal{..} -> do
                                     (bftx@BF.Transaction{_transactionIndex}, tx) <-
@@ -671,7 +636,7 @@ withNetworkLayer tr network np project k = do
             void $ LN.lightSync (MsgLightLayerLog >$< tr) lightSyncSource follower
 
     syncProgress :: BF.ClientConfig -> SlotNo -> IO SyncProgress
-    syncProgress bfConfig s = runBFM bfConfig do
+    syncProgress bfConfig s = BFM.run bfConfig do
         BF.Block {_blockSlot} <- BF.getLatestBlock
         let latestSlot = maybe 0 BF.unSlot _blockSlot
             currentSlot = fromIntegral (unSlotNo s)
@@ -681,7 +646,7 @@ withNetworkLayer tr network np project k = do
     stakePoolsSummary :: BF.ClientConfig -> Coin -> IO StakePoolsSummary
     stakePoolsSummary bfConfig _coin = do
         protocolParameters <- currentProtocolParameters bfConfig
-        runBFM bfConfig do
+        BFM.run bfConfig do
             BF.Network{_networkStake = BF.NetworkStake{_stakeLive}} <-
                 BF.getNetworkInfo
             totalLiveStake <- fromBlockfrostM _stakeLive
@@ -1174,40 +1139,3 @@ getPoolPerformanceEstimate sp dl rp pid = do
                     / fromIntegral (unCoin $ totalStake rp)
                     -- _poolHistoryActiveSize would be incorrect here
             }
-
-newtype BFM a = BFM (ReaderT BF.ClientConfig (ExceptT BlockfrostError IO) a)
-    deriving newtype
-        ( Functor
-        , Applicative
-        , Monad
-        , MonadIO
-        , MonadBase IO
-        , MonadBaseControl IO
-        , MonadReader BF.ClientConfig
-        , MonadError BlockfrostError
-        )
-
-instance BF.MonadBlockfrost BFM where
-    getConf = ask
-    liftBlockfrostClient act = BFM do
-        env <- asks fst
-        liftIO (runClientM act env) >>= either (throwError . ClientError) pure
-
-newClientConfig :: BF.Project -> IO BF.ClientConfig
-newClientConfig prj = (,prj) <$> BF.newEnvByProject prj
-
-runBFM :: BF.ClientConfig -> BFM a -> IO a
-runBFM cfg (BFM c) = handleBlockfrostError (runReaderT c cfg)
-
-handleBlockfrostError :: ExceptT BlockfrostError IO a -> IO a
-handleBlockfrostError =
-    either (throwIO . BlockfrostException) pure <=< runExceptT
-
-maybe404 :: BFM a -> BFM (Maybe a)
-maybe404 bfm = (Just <$> bfm) `catchError` \case
-    ClientError (Servant.FailureResponse _ (Servant.Response s _ _ _))
-        | s == status404 -> pure Nothing
-    e -> throwError e
-
-empty404 :: Monoid a => BFM a -> BFM a
-empty404 = (fromMaybe mempty <$>) . maybe404
